added sleeve detection
