+++ build.common.gradle
@@ -229,14 +229,14 @@ android {
             signingConfig signingConfigs.release
 
             ndk {
-                abiFilters "armeabi-v7a" /*, "arm64-v8a" */
+                abiFilters "armeabi-v7a" , "arm64-v8a"
             }
         }
         debug {
             debuggable true
             jniDebuggable true
             ndk {
-                abiFilters "armeabi-v7a" /*, "arm64-v8a" */
+                abiFilters "armeabi-v7a" , "arm64-v8a"
             }
         }
     }
